images/events-archiver/requirements.txt
google-cloud-logging==1.8.0
python-dateutil==2.7.5
google-cloud-storage==1.13.0